+++ src/viur_cli/scriptor/__init__.py
@@ -1,1 +1,1 @@
-from .cli import script
+from .cli import scriptor