+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This file documents any relevant changes.
 ## [1.0.0] 2023-03-29
 - feat: added `viur install` subcommand to install scriptor and vi
 - fix: `viur vi` command is deprecated use `viur install vi` instead
-- feat: added `viur script` command to use scriptor scripts
+- feat: added `viur scriptor` command to use scriptor scripts
 - fix: 2-to-3 script viur 3.3.4 updates
 - fix: unified version schema
 - feat: added `viur build release` command which build all flare and npm apps
@@ -13,6 +13,7 @@ This file documents any relevant changes.
 - feat: apps are now listed in a builds object in the project.json
 - feat: 2 app kinds flare and npm are supported
 - feat: added `viur project scanNpm` to add existing npm apps
+- feat: added `viur script` call different viur related scripts like 2to3, flare or get_pyodide
 
 ## [0.7.8] - 2023-03-27
 - chore: updated app_server to 0.9.3